Extract promote job modals to be used in the editor
